pyproject.toml
[project]
name = "ordigi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.8",
    "fabric>=3.2.2",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "paramiko>=3.5.1",
    "pydantic>=2.11.1",
    "pyyaml>=6.0.2",
    "tabulate>=0.9.0",
    "uvicorn>=0.34.0",
    "websockets>=15.0.1",
    "python-multipart>=0.0.20",
    "dotenv>=0.9.9",
]

[project.scripts]
ordigi = "ordigi.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
]
